+++ workflow/src/main/java/org/apache/oodt/cas/workflow/instrepo/MemoryWorkflowInstanceRepositoryFactory.java
@@ -38,8 +38,7 @@ public class MemoryWorkflowInstanceRepositoryFactory implements
      */
     public MemoryWorkflowInstanceRepositoryFactory() {
         pageSize = Integer.getInteger(
-                "org.apache.oodt.cas.workflow.instanceRep.pageSize", 20)
-                .intValue();
+            "org.apache.oodt.cas.workflow.instanceRep.pageSize", 20);
     }
 
     /*
